+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The official release will have a defined and more stable API. If you are already
   * UI support for some kinds of metadata: 
     * `:human-readable-build-label`
   * Added events `:pipeline-started` and `:pipeline-finished` (#155)
+  * Added `lambdacd.presentation.pipeline-structure/flatten-pipeline-representation` and `lambdacd.presentation.pipeline-structure/step-display-representation-by-step-id` to simplify getting information about a specific step
 * Bug fixes:
   * Catch Exception instead of Throwable in build steps to avoid catching Errors which cannot be handled (#148), thanks @hgsy!
 * Deprecated: